add Mistral and Cohere, pump version to 1.0.2

+++ sdk/src/main/java/io/github/locxngo/chroma/embedding/voyageai/TextEmbeddingModelResponse.java
@@ -46,6 +46,6 @@ public class TextEmbeddingModelResponse<N> extends PrintableObject {
     public static class IntDtype extends TextEmbeddingModelResponse<Integer> {
     }
 
-    public static class ByteDtype extends TextEmbeddingModelResponse<Byte> {
+    public static class ByteDtype extends TextEmbeddingModelResponse<Integer> {
     }
 }
